choose photos & supprt decimal number

+++ app/src/main/java/com/example/depository_system/frontInforms/FrontRukuInform.java
@@ -14,7 +14,7 @@ public class FrontRukuInform {
 
     public String materialType;
 
-    public int materialNum;
+    public double materialNum;
 
     public String materialUnit;
